requirements.txt
h5py==3.8.0
imageio==2.27.0
imagesize==1.4.1
imgaug==0.4.0
mmcv-full==1.6.0
mmengine==0.7.3
mmpycocotools==12.0.3
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
Pillow==9.5.0
pytorch-metric-learning==2.1.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
timm==0.5.4
lensless==1.0.4
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.11.4
torchvision==0.13.1
tqdm==4.65.0
waveprop==0.0.4
pycsou==1.0.6
yapf==0.30.0
-e .